+++ src/ui/panes/albums.rs
@@ -7,34 +7,47 @@ use crate::{
         mpd_client::{Filter, MpdClient, Tag},
     },
     ui::{
+        browser::BrowserPane,
         utils::dirstack::{DirStack, DirStackItem},
         widgets::browser::Browser,
         KeyHandleResultInternal,
     },
-    utils::macros::{status_info, status_warn},
+    utils::{
+        macros::{status_info, status_warn},
+        mouse_event::MouseEvent,
+    },
 };
 
-use super::{browser::DirOrSong, BrowserPane, Pane};
+use super::{browser::DirOrSong, Pane};
 use anyhow::{anyhow, Context, Result};
 use crossterm::event::KeyEvent;
 use itertools::Itertools;
-use ratatui::{prelude::Rect, widgets::ListItem, Frame};
+use ratatui::{
+    prelude::Rect,
+    widgets::{ListItem, StatefulWidget},
+    Frame,
+};
 
-#[derive(Debug, Default)]
+#[derive(Debug)]
 pub struct AlbumsPane {
     stack: DirStack<DirOrSong>,
     filter_input_mode: bool,
+    browser: Browser<DirOrSong>,
+}
+
+impl AlbumsPane {
+    pub fn new(context: &AppContext) -> Self {
+        Self {
+            stack: DirStack::default(),
+            filter_input_mode: false,
+            browser: Browser::new(context.config),
+        }
+    }
 }
 
 impl Pane for AlbumsPane {
-    fn render(&mut self, frame: &mut Frame, area: Rect, AppContext { config, .. }: &AppContext) -> Result<()> {
-        frame.render_stateful_widget(
-            Browser::new(config)
-                .set_widths(&config.theme.column_widths)
-                .set_border_style(config.as_border_style()),
-            area,
-            &mut self.stack,
-        );
+    fn render(&mut self, frame: &mut Frame, area: Rect, _context: &AppContext) -> Result<()> {
+        self.browser.render(area, frame.buffer_mut(), &mut self.stack);
 
         Ok(())
     }
@@ -88,6 +101,15 @@ impl Pane for AlbumsPane {
             }
             _ => Ok(KeyHandleResultInternal::SkipRender),
         }
+    }
+
+    fn handle_mouse_event(
+        &mut self,
+        event: MouseEvent,
+        client: &mut impl MpdClient,
+        context: &mut AppContext,
+    ) -> Result<KeyHandleResultInternal> {
+        self.handle_mouse_action(event, client, context)
     }
 
     fn handle_action(
@@ -242,5 +264,9 @@ impl BrowserPane<DirOrSong> for AlbumsPane {
                     _ => None,
                 })
             })
+    }
+
+    fn browser_areas(&self) -> [Rect; 3] {
+        self.browser.areas
     }
 }