+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
   - changing active tab
   - changing focused pane in tab splits
   - scrolling in queue, clicking on table to select song, double click to play song
+  - browser panes, double click to enter dir/add song, scrolling, click next/previous area to advance/return
 
 ### Changed
 
@@ -22,6 +23,7 @@ All notable changes to this project will be documented in this file.
 - Rmpc now logs warnings and errors in CLI mode to stderr
 - try to clean up after yt-dlp in case it fails
 - Album art not clearing properly after direct tab switch
+- Events being duplicated when panes were present in multiple tabs
 
 ## [0.5.0] - 2024-09-27